Update to 0.8.0
This manifest was generated with
https://hackage.haskell.org/package/cabal-flatpak in the following
rather convoluted way:

0. Patch cabal-flatpak to always include alex and happy in the build
   plan
1. Install cabal-flatpak (on the host system)
2. Generate a build plan for bustle-0.8.0:

    cd ~/tmp
    cabal unpack bustle-0.8.0
    cd bustle-0.8.0
    cabal new-build --dry-run --disable-tests --disable-benchmarks --flags=-hgettext

3. Run cabal-flatpak against that plan
4. Build the resulting manifest with --stop-at=libpcap
5. Enter the build environment with `flatpak-builder --run _build/ flatpak/org.freedesktop.Bustle.json bash`
6. Generate a new build plan within that build environment
7. Re-run cabal-flatpak
8. Manually edit the resulting manifest to add a final stanza to run make

Step 0 is needed because gtk2hs has a build-tool dependency on alex and
happy executables, but cabal-flatpak upstream only includes library
dependencies.

Steps 4-7 are needed because cabal on the host system will use whatever
library versions happen to be installed as part of its build plan. These
may conflict with the library versions present in the GHC distribution
used by cabal-flatpak. On my system the result failed to build.

Step 8 is needed because the Cabal-based build system does not install
appdata, icons, ...